backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.7.4
python-multipart==0.0.6
pymongo==4.6.1
python-dotenv==1.0.0
cachetools==5.3.2
zhipuai==2.0.1
# 文件处理依赖
PyMuPDF==1.23.8
python-docx==0.8.11
Pillow==10.0.0
pytesseract==0.3.10
# 阿里云OSS SDK
oss2==2.18.4
